requirements.txt
pyyaml
confluent-kafka
pyarrow
pandas
